requirements.txt
streamlit
opencv-python-headless
numpy
onnxruntime
rembg
Pillow
mahotas